Upgrade WPUnit tests to Integration Tests
