+++ tests/Integration/_bootstrap.php
@@ -1,0 +1,1 @@
+<?php
