--- tests/unit.suite.yml
@@ -1,10 +1,0 @@
-# Codeception Test Suite Configuration
-#
-# Suite for unit tests not relying WordPress code.
-
-actor: UnitTester
-modules:
-    enabled:
-        - Asserts
-        - \Helper\Unit
-    step_decorators: ~
