--- tests/wpunit.suite.yml
@@ -1,21 +1,0 @@
-# Codeception Test Suite Configuration
-#
-# Suite for unit or integration tests that require WordPress functions and classes.
-
-actor: WpunitTester
-modules:
-    enabled:
-        - WPLoader
-        - \Helper\Wpunit
-    config:
-        WPLoader:
-            wpRootFolder: "%WP_ROOT_FOLDER%"
-            dbName: "%TEST_DB_NAME%"
-            dbHost: "%TEST_DB_HOST%"
-            dbUser: "%TEST_DB_USER%"
-            dbPassword: "%TEST_DB_PASSWORD%"
-            tablePrefix: "%TEST_TABLE_PREFIX%"
-            domain: "%TEST_SITE_WP_DOMAIN%"
-            adminEmail: "%TEST_SITE_ADMIN_EMAIL%"
-            title: "Test"
-            plugins: ['convertkit/wp-convertkit.php'] # Change to the repository Plugin that we're testing.
